+++ src/main/java/com/karankumar/bookproject/ui/book/BookForm.java
@@ -28,8 +28,8 @@ import com.karankumar.bookproject.backend.service.PredefinedShelfService;
 import com.karankumar.bookproject.backend.utils.CustomShelfUtils;
 import com.karankumar.bookproject.backend.utils.PredefinedShelfUtils;
 import com.karankumar.bookproject.ui.book.components.BookGenreComboBox;
-import com.karankumar.bookproject.ui.book.components.PagesReadFormItem;
-import com.karankumar.bookproject.ui.book.components.RatingFormItem;
+import com.karankumar.bookproject.ui.book.components.form.item.PagesReadFormItem;
+import com.karankumar.bookproject.ui.book.components.form.item.RatingFormItem;
 import com.karankumar.bookproject.ui.components.utils.ComponentUtil;
 import com.vaadin.flow.component.ComponentEvent;
 import com.vaadin.flow.component.ComponentEventListener;
